refactor: eliminate duplicative CLI structure and consolidate command logic
- Remove duplicate core/init.py and core/update.py files
- Create shared utilities: core/detector.py (RepositoryDetector) and core/deployer.py (TemplateDeployer, ParameterSubstitutor)
- Consolidate all init logic in commands/init.py
- Consolidate all update logic in commands/update.py
- Establish clean separation: core/ for shared utilities, commands/ for command-specific logic
- Update tests to reflect new architecture

This architectural refactoring eliminates confusion between core/ and commands/
directories while maintaining all functionality. (refs #205)
